Move test functions to correct module
